+++ webapp/src/app/components/dialogs/settings/settings.component.ts
@@ -6,9 +6,10 @@ import { BrowserService } from '../../../services/browser.service';
 import { ElectronService } from '../../../services/electron.service';
 import { Globals } from '../../../services/globals';
 import { HttpClientService } from '../../../services/http-client.service';
-import { SettingsService } from '../../../services/settings.service';
+import { AppSettings, SettingsService } from '../../../services/settings.service';
 import { SharedService } from '../../../services/shared-service';
 import { OverlayRefControl } from '../overlay/overlay-ref-control';
+import { PropListCard } from '../../widgets/shared/image-select-overlay/image-select-card/image-select-card.component';
 
 @Component({
 	selector: 'app-settings',
@@ -16,17 +17,28 @@ import { OverlayRefControl } from '../overlay/overlay-ref-control';
 	styleUrls: ['./settings.component.scss']
 })
 export class SettingsComponent implements OnInit {
-
+	
 	isElectron = Globals.isElectron;
 	folderFormControl = new FormControl();
 	icon = 'help_outline';
 	iconCss = 'icon-undefined';
 	mods: string[] = [];
 	mod = '';
-	wrapEventEditorLines: boolean;
-
+	settings: AppSettings;
+	isIncludeVanillaMapsDisabled: boolean;
+	
+	cardLight: PropListCard = {
+		name: 'Light',
+		imgSrc: 'assets/selection-light.png',
+	};
+	
+	cardDark: PropListCard = {
+		name: 'Dark',
+		imgSrc: 'assets/selection-dark.png',
+	};
+	
 	private readonly sharedService: SharedService;
-
+	
 	constructor(
 		private ref: OverlayRefControl,
 		private electron: ElectronService,
@@ -40,26 +52,27 @@ export class SettingsComponent implements OnInit {
 		} else {
 			this.sharedService = browser;
 		}
-
+		
 		http.getMods().subscribe(mods => this.mods = mods);
 		this.mod = this.sharedService.getSelectedMod();
-		this.wrapEventEditorLines = this.settingsService.wrapEventEditorLines;
+		this.isIncludeVanillaMapsDisabled = !this.mod;
+		this.settings = JSON.parse(JSON.stringify(this.settingsService.getSettings()));
 	}
-
+	
 	ngOnInit() {
 		if (this.isElectron) {
 			this.folderFormControl.setValue(this.electron.getAssetsPath());
 			this.folderFormControl.valueChanges.subscribe(() => this.resetIcon());
 		}
-
+		
 		this.check();
 	}
-
+	
 	private resetIcon() {
 		this.icon = 'help_outline';
 		this.iconCss = 'icon-undefined';
 	}
-
+	
 	private setIcon(valid: boolean) {
 		if (valid) {
 			this.icon = 'check';
@@ -69,14 +82,14 @@ export class SettingsComponent implements OnInit {
 			this.iconCss = 'icon-invalid';
 		}
 	}
-
+	
 	select() {
 		const path = this.electron.selectCcFolder();
 		if (path) {
 			this.folderFormControl.setValue(path);
 		}
 	}
-
+	
 	check() {
 		const valid = this.electron.checkAssetsPath(this.folderFormControl.value);
 		this.setIcon(valid);
@@ -88,23 +101,27 @@ export class SettingsComponent implements OnInit {
 			});
 		}
 	}
-
+	
+	modSelectEvent(selectedMod: string) {
+		this.isIncludeVanillaMapsDisabled = !selectedMod;
+	}
+	
 	save() {
 		if (this.isElectron) {
 			this.electron.saveAssetsPath(this.folderFormControl.value);
 		}
 		this.sharedService.saveModSelect(this.mod);
-		this.settingsService.wrapEventEditorLines = this.wrapEventEditorLines;
+		this.settingsService.updateSettings(this.settings);
 		this.close();
 		const ref = this.snackBar.open('Changing the path requires to restart the editor', 'Restart', {
 			duration: 6000
 		});
-
+		
 		ref.onAction().subscribe(() => this.sharedService.relaunch());
 	}
-
+	
 	close() {
 		this.ref.close();
 	}
-
+	
 }