+++ webapp/src/app/services/phaser/tilemap/tile-drawer.ts
@@ -15,7 +15,7 @@ export class TileDrawer extends BaseObject {
 	private layer?: CCMapLayer;
 	private selectedTiles: SelectedTile[] = [];
 	
-	private rect?: Phaser.GameObjects.Rectangle;
+	private selection?: Phaser.GameObjects.Container;
 	
 	private previewTileMap!: Phaser.Tilemaps.Tilemap;
 	private previewLayer?: Phaser.Tilemaps.TilemapLayer;
@@ -99,7 +99,7 @@ export class TileDrawer extends BaseObject {
 			return;
 		}
 		const pointer = this.scene.input.activePointer;
-		const p = Helper.worldToTile(pointer.worldX, pointer.worldY);
+		const p = Helper.worldToTile(pointer.worldX - this.layer.x, pointer.worldY - this.layer.y);
 		
 		// render selection border
 		if (this.rightClickStart) {
@@ -126,7 +126,7 @@ export class TileDrawer extends BaseObject {
 				diff.y--;
 			}
 			
-			this.drawRect(diff.x, diff.y, start.x, start.y);
+			this.drawRect(diff.x, diff.y, start.x, start.y, true);
 			return;
 		}
 		
@@ -135,15 +135,15 @@ export class TileDrawer extends BaseObject {
 		container.x = pointer.worldX;
 		container.y = pointer.worldY;
 		
-		if (container.x < 0) {
+		if (container.x < this.layer.x) {
 			container.x -= Globals.TILE_SIZE;
 		}
-		if (container.y < 0) {
+		if (container.y < this.layer.y) {
 			container.y -= Globals.TILE_SIZE;
 		}
 		
-		container.x -= container.x % Globals.TILE_SIZE;
-		container.y -= container.y % Globals.TILE_SIZE;
+		container.x -= (container.x - this.layer.x) % Globals.TILE_SIZE;
+		container.y -= (container.y - this.layer.y) % Globals.TILE_SIZE;
 		
 		if (this.previewLayer) {
 			Vec2.assign(this.previewLayer, container);
@@ -151,7 +151,7 @@ export class TileDrawer extends BaseObject {
 		
 		// draw tiles
 		// trigger only when mouse is over canvas element (the renderer), avoids triggering when interacting with ui
-		if (pointer.leftButtonDown() && pointer.downElement.nodeName === 'CANVAS' && this.layer) {
+		if (pointer.leftButtonDown() && pointer.downElement?.nodeName === 'CANVAS' && this.layer) {
 			const finalPos = {x: 0, y: 0};
 			const startPos = {x: 0, y: 0};
 			
@@ -297,7 +297,7 @@ export class TileDrawer extends BaseObject {
 		
 		// only start tile copy when cursor in bounds
 		const pointer = this.scene.input.activePointer;
-		const p = Helper.worldToTile(pointer.worldX, pointer.worldY);
+		const p = Helper.worldToTile(pointer.worldX - this.layer.x, pointer.worldY - this.layer.y);
 		if (!Helper.isInBounds(this.layer, p)) {
 			return;
 		}
@@ -307,15 +307,64 @@ export class TileDrawer extends BaseObject {
 		this.rightClickStart = p;
 	}
 	
-	private drawRect(width: number, height: number, x = 0, y = 0) {
-		if (this.rect) {
-			this.rect.destroy();
+	private drawRect(width: number, height: number, x = 0, y = 0, renderSize = false) {
+		if (this.selection) {
+			this.selection.destroy();
 		}
-		this.rect = this.scene.add.rectangle(x, y, width * Globals.TILE_SIZE, height * Globals.TILE_SIZE);
-		this.rect.setOrigin(0, 0);
-		this.rect.setStrokeStyle(1, 0xffffff, 0.6);
 		
-		this.container.add(this.rect);
+		let textColor = 'rgba(0,0,0,0.6)';
+		let backgroundColor = 0xffffff;
+		if (Globals.settingsService.getSettings().selectionBoxDark) {
+			textColor = 'rgba(255,255,255,0.9)';
+			backgroundColor = 0x333333;
+		}
+		
+		this.selection = this.scene.add.container(x, y);
+		
+		const rect = this.scene.add.rectangle(0, 0, width * Globals.TILE_SIZE, height * Globals.TILE_SIZE);
+		rect.setOrigin(0, 0);
+		rect.setStrokeStyle(1, backgroundColor, 0.6);
+		
+		this.selection.add(rect);
+		this.container.add(this.selection);
+		
+		if (!renderSize) {
+			Globals.globalEventsService.updateTileSelectionSize.next(undefined);
+			return;
+		}
+		
+		const makeText = (pos: Point, val: number) => {
+			const text = this.scene.add.text(pos.x, pos.y, Math.abs(val) + '', {
+				font: '400 10px Roboto',
+				color: textColor,
+				resolution: window.devicePixelRatio * 3,
+			});
+			text.setOrigin(0.5, 0);
+			const background = this.scene.add.rectangle(pos.x, pos.y + 2, 14, 10, backgroundColor, 0.6);
+			background.setOrigin(0.5, 0);
+			
+			this.selection?.add(background);
+			this.selection?.add(text);
+		};
+		
+		if (Math.abs(width) >= 3) {
+			makeText({
+				x: width * Globals.TILE_SIZE / 2,
+				y: (height > 0 ? 0 : height * Globals.TILE_SIZE) - 1
+			}, width);
+		}
+		
+		if (Math.abs(height) >= 3) {
+			makeText({
+				x: Globals.TILE_SIZE / 2 + (width > 0 ? 0 : width * Globals.TILE_SIZE),
+				y: (height - 1) * Globals.TILE_SIZE / 2,
+			}, height);
+		}
+		
+		Globals.globalEventsService.updateTileSelectionSize.next({
+			x: Math.abs(width),
+			y: Math.abs(height)
+		});
 	}
 	
 	private onMouseRightUp() {
@@ -398,10 +447,14 @@ export class TileDrawer extends BaseObject {
 		}
 		
 		const pointer = this.scene.input.activePointer;
-		const p = Helper.worldToTile(pointer.worldX, pointer.worldY);
+		const p = Helper.worldToTile(pointer.worldX - this.layer.x, pointer.worldY - this.layer.y);
 		
 		if (this.selectedTiles.length > 0) {
 			Filler.fill(this.layer, this.selectedTiles[0].id, p);
+			Globals.stateHistoryService.saveState({
+				name: 'fill',
+				icon: 'format_color_fill'
+			});
 		}
 		
 	}